# OPT: Speedup of per-subdataset "contains?" matching