--- a/datalad/local/tests/test_subdataset.py
+++ b/datalad/local/tests/test_subdataset.py
@@ -192,8 +192,8 @@ def test_get_subdatasets(origpath, path):
     res = ds.subdatasets(recursive=True)
     assert_status('ok', res)
     for r in res:
-        #for prop in ('gitmodule_url', 'state', 'revision', 'gitmodule_name'):
-        for prop in ('gitmodule_url', 'revision', 'gitmodule_name'):
+        #for prop in ('gitmodule_url', 'state', 'gitshasum', 'gitmodule_name'):
+        for prop in ('gitmodule_url', 'gitshasum', 'gitmodule_name'):
             assert_in(prop, r)
         # random property is unknown
         assert_not_in('mike', r)