Speeds up the rejection of non-containing subdatasets. For the UKB dataset with 42k subdatasets, this shaves of 1s of the previous total runtime of 2.5s.

Related to gh-4859

This also removes a backward compatibility kludge that we put in place over a year ago (`revision` instead of the present `gitshasum` property in the submodule records).

Edit: This also ups the plain reporting performance (same 42k subdatasets):

Before: `datalad subdatasets  9.49s user 3.41s system 57% cpu 22.284 total`
After: `datalad subdatasets  9.81s user 2.49s system 74% cpu 16.399 total`
and with a twist: `datalad subdatasets -d .  6.95s user 2.14s system 66% cpu 13.622 total` (see TODO)

TODO:

- [x] `GitRepo.get_submodules_()` calls `GitRepo.get_content_info()`. This might be made faster, by filtering the path-contraints with the `contains` parameter before calling any of this stack. Right now `contains` is considered last and all info is requested without any path constraint in the context of #4859 -> #5063 
- [x] Suprise differences with or without `--dataset`
    ```
    datalad subdatasets -r --contains sub-3507513  2.76s user 0.33s system 103% cpu 2.973 total                                           
    datalad subdatasets -d . -r --contains sub-3507513  1.36s user 0.21s system 104% cpu 1.507 total                                      
    ```
- [x] `update` is broken by the second commit, not clear why yet (likely it is still using the `revision` property)