@yarikoptic 2020-09-04
FWIW -- no significant impact on existing benchmarks (if anything - slight tendency to slow down).  Would be nice to get a dedicated benchmark to demonstrate the effect.