@mih 2020-09-04
> FWIW -- no significant impact on existing benchmarks (if anything - slight tendency to slow down). Would be nice to get a dedicated benchmark to demonstrate the effect.

Can you clarify what you want? From the times you see above you can decide on how big of a dataset you need to show an effect. For 1s you need 40k subdatasets, if you are satisfied with less, a few thousand should work. So let's say you want to see 100ms and do 4k subdatasets, you need to invest the ~3h to create such a dataset.