@yarikoptic 2020-09-04
are you saying that there should be no notable (but thus also no negative) impact on a superdataset with e.g. 20 subdatasets?

Moreover we already have some benchmarks setup with some sizeable (although may be not big engouh) datasets, e.g. subclasses of https://github.com/datalad/datalad/blob/master/benchmarks/common.py#L104  - might be worth adding a  `time_contains` in one of those subclasses.