@mih 2020-09-04
There should be no negative impact on small size datasets.