@adswa 2020-09-07
Here is what I get (on NFS):

In master:

```
datalad subdatasets > somelog 11.98s user 1.08s system 100% cpu 13.023 total
datalad subdatasets -d . > somelog 10.05s user 1.15s system 100% cpu 11.179 total
datalad subdatasets > /dev/null  12.45s user 2.10s system 72% cpu 20.132 total   (???)
datalad subdatasets -d . > /dev/null  9.14s user 0.90s system 100% cpu 10.022 total
datalad -f '{path}' subdatasets > /dev/null  10.39s user 0.90s system 100% cpu 11.241 total
datalad -f '{path}' subdatasets -d . > /dev/null  7.19s user 0.91s system 100% cpu 8.078 total
```

On this branch:

```
SLOWER datalad subdatasets > somelog 11.06s user 1.87s system 76% cpu 16.922 total
FASTER datalad subdatasets -d . > somelog  8.62s user 0.87s system 100% cpu 9.416 total
FASTER datalad subdatasets > /dev/null   11.49s user 1.90s system 77% cpu 17.305 total
FASTER datalad subdatasets -d . > /dev/null   8.61s user 0.84s system 100% cpu 9.427 total
SLOWER datalad -f '{path}' subdatasets > /dev/null  9.09s user 1.95s system 73% cpu 15.015 total
FASTER datalad -f '{path}' subdatasets -d . > /dev/null 6.83s user 0.87s system 100% cpu 7.695 total
```

@mih's stats in comparison (on local drive):

```
datalad subdatasets > somelog  11.18s user 2.10s system 76% cpu 17.326 total
datalad subdatasets -d . > somelog  8.42s user 0.74s system 100% cpu 9.102 total
datalad subdatasets > /dev/null  11.00s user 1.17s system 101% cpu 12.001 total
datalad subdatasets -d . > /dev/null  8.69s user 2.04s system 69% cpu 15.361 total
datalad -f '{path}' subdatasets > /dev/null  8.87s user 1.01s system 101% cpu 9.714 total
datalad -f '{path}' subdatasets -d . > /dev/null  6.55s user 1.84s system 65% cpu 12.799 total
```