@yarikoptic 2020-09-07
Wild guess re `-f '{path}'` effect - may be there is some `.repo` is involved in default full renderer?