@mih 2020-10-21
No different conditional on `--dataset` anymore, with or without writing to a log:

```
datalad subdatasets > somelog  7.45s user 2.03s system 100% cpu 9.395 total
datalad subdatasets -d . > somelog  7.57s user 1.94s system 100% cpu 9.434 total
datalad subdatasets > /dev/null  7.48s user 1.98s system 100% cpu 9.396 total
datalad subdatasets -d . > /dev/null  7.44s user 2.10s system 100% cpu 9.477 total
```

Still a substantial impact of the result rendering:

```
datalad -f '{path}' subdatasets > /dev/null  6.19s user 2.10s system 100% cpu 8.219 total
```

But there is nothing `subdatasets`  specific about this, hence not subject of this PR.