@mih 2020-10-21
Current benchmark runs seem to indicate no performance penalty for the small scale dataset we are using in the tests

https://github.com/datalad/datalad/pull/4868/checks?check_run_id=1286049657