@mih 2020-10-21
FTR: I tried to further improve the situation by using the `--contains` specification as additional query constraints, but that did not work -- made it slower. For posterity, here was my patch:

```diff
diff --git a/datalad/local/subdatasets.py b/datalad/local/subdatasets.py
index cc36a074f..5c0d63b9d 100644
--- a/datalad/local/subdatasets.py
+++ b/datalad/local/subdatasets.py
@@ -75,6 +75,14 @@ def _parse_git_submodules(ds_pathobj, repo, paths):
             else:
                 # we had path contraints, but none matched this dataset
                 return
+        else:
+            # filter out all the paths that point outside the repo. they would
+            # throw errors in GitRepo.get_submodules_(). such paths would
+            # make it here, because we mix top-level query path constrains with
+            # --contains specification in order to speed things up. If the latter
+            # are garbage, not reporting on them will be detected and impossible
+            # results are generated
+            paths = [p for p in paths if p.parts and p.parts[0] != os.pardir]
     # can we use the reported as such, or do we need to recode wrt to the
     # query context dataset?
     if ds_pathobj == repo.pathobj:
@@ -289,8 +297,12 @@ def _get_submodules(ds, paths, fulfilled, recursive, recursion_limit,
         [c] + list(c.parents)
         for c in (contains if contains is not None else [])
     ]
+    submodule_path_constraints = (paths or []) + contains if contains else paths
     # put in giant for-loop to be able to yield results before completion
-    for sm in _parse_git_submodules(ds.pathobj, repo, paths):
+    for sm in _parse_git_submodules(
+            ds.pathobj,
+            repo,
+            submodule_path_constraints):
         sm_path = sm['path']
         contains_hits = None
         if contains:
```